requirements.txt
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
click==8.1.7
easyocr==1.7.1
einops==0.8.0
filelock==3.15.4
packaging==24.1
huggingface-hub==0.24.3
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.1
more-itertools==8.10.0
mpmath==1.3.0
netifaces==0.11.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
pandas==2.2.2
pillow==10.4.0
psutil==6.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post5
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pyspellchecker==0.8.1
pytz==2024.1
fastapi
scikit-image==0.24.0
scipy==1.14.0
SecretStorage==3.3.1
shapely==2.0.5
six==1.16.0
sympy==1.13.1
tifffile==2024.7.24
timm==1.0.8
tokenizers==0.19.1
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.4
transformers==4.43.3
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.70
ultralytics-thop==2.0.0
urllib3==2.2.2
uvicorn==0.30.3
wadllib==1.3.6
zipp==1.0.0
